authentication and syntax suger

+++ Cairo_book_fair/Controllers/CartController.cs
@@ -5,9 +5,11 @@ using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using System.Net;
+using System.Security.Claims;
 
 namespace Cairo_book_fair.Controllers
 {
+    [Authorize]
     [Route("api/[controller]")]
     [ApiController]
     public class CartController : ControllerBase
@@ -20,18 +22,23 @@ namespace Cairo_book_fair.Controllers
             _bookCartService = bookCartService;
             _cartService = cartService;
         }
-
         [HttpGet("All")]
-        public IActionResult GetAllItems(int cartId)
+        public IActionResult GetAllItems(string UserId)
         {
-            List<CartItemDTO> cartItemsDto = _bookCartService.GetAllCartItems(cartId);
-
-            if(cartItemsDto == null)
+            if(UserId == User.FindFirstValue(ClaimTypes.NameIdentifier)) 
             {
-                return BadRequest();
+                Cart cart = _cartService.GetCartByUserId(UserId);
+                WholeCartItemsWithTotalPriceDTO Cart = _bookCartService.GetAllCartItems(cart.Id);
+
+                if(Cart.cartItems.Count == 0)
+                {
+                    return BadRequest("!نأسف لعدم وجود اي كتاب في السلة .. اشتري الآن");
+                }
+
+                return Ok(Cart);
             }
 
-            return Ok(cartItemsDto);
+            return Unauthorized("عذراً هذا الحساب ليس له صلاحية الوصول للسلة");
         }
 
         //[HttpGet]
@@ -39,11 +46,10 @@ namespace Cairo_book_fair.Controllers
         //{
 
         //}
-
         [HttpPost ("Buy Regular Book")]
         public IActionResult InserItem(BookItemWithUserID bookItemWithUserID)
         {
-            if(ModelState.IsValid)
+            if(ModelState.IsValid && bookItemWithUserID.userId == User.FindFirstValue(ClaimTypes.NameIdentifier))
             {
                  Cart cart = _cartService.GetCartByUserId(bookItemWithUserID.userId);
                 _bookCartService.AddItem(cart.Id, bookItemWithUserID.bookId);
@@ -53,7 +59,6 @@ namespace Cairo_book_fair.Controllers
             }
             return BadRequest("عذراً لم يتم اضافة الكتاب إلى السلة");
         }
-
         [HttpPost("Take Donated Book")]
         public IActionResult TakeBook(BookItemWithUserID bookItemWithUserID)
         {
@@ -67,34 +72,35 @@ namespace Cairo_book_fair.Controllers
             }
             return BadRequest("عذراً لم يتم اضافة الكتاب إلى السلة");
         }
-
-        [HttpDelete]
+        
+        [HttpDelete("Remove-item-from-Cart")]
         public IActionResult Delete(BookItemWithUserID bookItemWithUserID)
         {
-            if (ModelState.IsValid)
+            if (ModelState.IsValid && bookItemWithUserID.userId == User.FindFirstValue(ClaimTypes.NameIdentifier))
             {
                 Cart cart = _cartService.GetCartByUserId(bookItemWithUserID.userId);
                 _bookCartService.RemoveItem(cart.Id, bookItemWithUserID.bookId);
                 _bookCartService.Save();
-                return NoContent();
+                return Ok("!تمت ازالة الكتاب بنجاح");
             }
 
-            return BadRequest(ModelState);          
+            return BadRequest("عذراً لم تتم عملية الازالة");
 
         }
 
         [HttpPut("Change Quantity")]
         public IActionResult ChangeQuantity(int quantity, BookItemWithUserID bookItemWithUserID)
         {
-            if (ModelState.IsValid)
+            if (ModelState.IsValid && bookItemWithUserID.userId == User.FindFirstValue(ClaimTypes.NameIdentifier))
+
             {
                 Cart cart = _cartService.GetCartByUserId(bookItemWithUserID.userId);
                 _bookCartService.RemoveItem(cart.Id, bookItemWithUserID.bookId);
                 _bookCartService.Save();
-                return NoContent();
+                return Ok("!تم تعديل الكمية بنجاح");
             }
 
-            return BadRequest(ModelState);
+            return BadRequest("عذراً لم تتم عملية التعديل");
 
         }
